Add knobs for sched features
Own implementation to play with Gentle Fair Sleepers and Arch Capacity.
It's a good time to say that nearly all commits cherry-picked from old implementation are totally broken. In 3.18 there's no anything like Arch Power as it was available in ~3.10!
A correspondent for it is Arch Capacity which provide *common* scaling or *arch depedent* scaling.

Signed-off-by: xNombre <kartapolska@gmail.com>
Signed-off-by: MOVZX <movzx@yahoo.com>