msm: camera_v2: fix for older blobs
It took me a few days to notice that size of this structure is too small.
It was breaking camera because ISP command wasn't recognized.

It's a dummy add of commit 3d5e5360d2aa9 "msm: camera: isp: Handle standby mode for B+M"
Enum is at size of int32.

There's no need to port it as standby mode seems to be implemented in other way.

Signed-off-by: xNombre <kartapolska@gmail.com>